change from jsx to tsx
